+++ include/datastructures/hash.h
@@ -19,15 +19,15 @@ typedef uint64_t((*hash_func)(const struct hash *h, const void *k));
 
 struct hash {
 	struct arr meta, e, keys;
-	size_t cap, len, load, max_load, capm;
+	uint32_t cap, len, load, max_load, capm;
 	hash_keycmp keycmp;
 	hash_func hash_func;
 };
 
 typedef enum iteration_result((*hash_with_keys_iterator_func)(void *ctx, const void *key, uint64_t val));
 
-void hash_init(struct hash *h, size_t cap, uint32_t keysize);
-void hash_init_str(struct hash *h, size_t cap);
+void hash_init(struct hash *h, uint32_t cap, uint32_t keysize);
+void hash_init_str(struct hash *h, uint32_t cap);
 void hash_destroy(struct hash *h);
 
 uint64_t *hash_get(const struct hash *h, const void *key);
